Move all tests and test data into tests/ directory
Tests are no longer in the same Python file as the implementation. This
makes it easier to split custodia into a client and server package.
