--- conftest.py
@@ -1,1 +1,0 @@
-collect_ignore = ["setup.py", "custodia/custodia"]
