+++ tox.ini
@@ -58,7 +58,6 @@ commands =
 
 [pytest]
 norecursedirs = build .tox
-python_files = tests/*.py custodia/*.py
 
 [flake8]
 exclude = .tox,*.egg,dist,build,docs/source
